+++ src/bootstrap/container.ts
@@ -15,6 +15,7 @@ import { GraphGenerationService } from '../core/services/GraphGenerationService'
 import { SyncService } from '../core/services/SyncService';
 import { NotificationService } from '../core/services/NotificationService';
 import { AiInteractionService } from '../core/services/AiInteractionService';
+import { AgentService } from '../core/services/AgentService';
 import { GoogleOAuthService } from '../core/services/GoogleOAuthService';
 import { AppleOAuthService } from '../core/services/AppleOAuthService';
 import { MicroscopeManagementService } from '../core/services/MicroscopeManagementService';
@@ -83,6 +84,7 @@ export class Container {
   private syncService: SyncService | null = null;
   private notificationService: NotificationService | null = null;
   private aiInteractionService: AiInteractionService | null = null;
+  private agentService: AgentService | null = null;
   private googleOAuthService: GoogleOAuthService | null = null;
   private appleOAuthService: AppleOAuthService | null = null;
   private microscopeManagementService: MicroscopeManagementService | null = null;
@@ -369,7 +371,23 @@ export class Container {
   }
 
   /**
-   *
+   * AgentService 인스턴스를 반환합니다.
+   */
+  getAgentService(): AgentService {
+    if (!this.agentService) {
+      const raw = new AgentService({
+        noteService: this.getNoteService(),
+        conversationService: this.getConversationService(),
+        messageService: this.getMessageService(),
+        graphEmbeddingService: this.getGraphEmbeddingService(),
+        graphVectorService: this.getGraphVectorService(),
+      });
+      this.agentService = createAuditProxy(raw, 'AgentService');
+    }
+    return this.agentService;
+  }
+
+  /**
    * AiInteractionService 인스턴스를 반환합니다.
    */
   getAiInteractionService(): AiInteractionService {
